requirements.txt
PyAutoGUI==0.9.54
pillow==11.0.0
opencv-python==4.10.0.84